FoodDrink, Review, Menu, Order Services were updated.
New properties were added. Errors were fixed.

+++ onlineFoodOrdering/src/main/java/com/onlineFoodOrdering/onlineFoodOrdering/service/ReviewService.java
@@ -8,6 +8,7 @@ import com.onlineFoodOrdering.onlineFoodOrdering.repository.CustomerRepository;
 import com.onlineFoodOrdering.onlineFoodOrdering.repository.OrderRepository;
 import com.onlineFoodOrdering.onlineFoodOrdering.repository.RestaurantRepository;
 import com.onlineFoodOrdering.onlineFoodOrdering.repository.ReviewRepository;
+import com.onlineFoodOrdering.onlineFoodOrdering.request.ReviewCreateRequest;
 import com.onlineFoodOrdering.onlineFoodOrdering.response.ReviewResponse;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -57,50 +58,51 @@ public class ReviewService {
 
     }
 
-    public String addAReview(Long id, Review review, Long restaurantId){
-        Customer customer = findCustomer(id);
+    public String addAReview(Long customerId, ReviewCreateRequest review, Long restaurantId){
+        Customer customer = findCustomer(customerId);
 
         Restaurant restaurant = restaurantRepository.findById(restaurantId).orElse(null);
-        Review existReview = reviewRepository.findReviewByCustomerIdAndRestaurantId(id,restaurantId).orElse(null);
+        //Review existReview = reviewRepository.findReviewByCustomerIdAndRestaurantId(customerId,restaurantId).orElse(null);
 
         // The purpose is that giving permission to a customer make comment a restaurant with a constraint
         // which is that he/she only make a comment for the restaurants which were ordered something by him/her.
-        List<Order> ordersForTheCustomer =  orderRepository.findOrdersByCustomerId(id);
+        List<Order> ordersForTheCustomer =  orderRepository.findOrdersByCustomerId(customerId);
 
-        boolean isReal = false;
+        boolean isOrderedBefore = false;
 
         for (int i = 0; i < ordersForTheCustomer.size(); i++) {
             if(ordersForTheCustomer.get(i).getRestaurant() == restaurant){
-                isReal = true;
+                isOrderedBefore = true;
                 break;
             }
         }
 
-        if (isReal) {
-            if(existReview == null){
-                if(customer == null && restaurant == null){
-                    return "Please enter an available id and restaurant.";
-                }
-                else if(customer == null){
-                    return "Please enter an available id.";
-                }
-                else if(restaurant == null){
-                    return "Please enter an available restaurant.";
-                }else{
+        if(customer == null && restaurant == null){
+            return "Please enter an available id and restaurant.";
+        }
+        else if(customer == null){
+            return "There is no such a customer.";
+        }
+        else if(restaurant == null){
+            return "Please enter an available restaurant.";
+        }
+
+
+
+        if (isOrderedBefore) {
+            //if(existReview == null){
                     Review newReview = new Review();
                     newReview.setCustomer(customer);
                     newReview.setText(review.getText());
                     newReview.setTitle(review.getTitle());
                     newReview.setPoint(review.getPoint());
-                    newReview.setRestaurant(review.getRestaurant());
-
+                    newReview.setRestaurant(restaurant);
                     reviewRepository.save(newReview);
 
                     return "Your review is saved successfully.";
-                }
-            }else{
-                return "You have already added a review for this restaurant.";
-            }
+            //}else{
+            //    return "You have already added a review for this restaurant.";
+            //}
         }else{
             return "You cannot make a review for a restaurant you did not order anything.";
         }
@@ -108,39 +110,43 @@ public class ReviewService {
 
     }
 
-    public String updateReview(Long id, Review review,Long restaurantId){
-        //Customer customer = findCustomer(id);
+    public String updateReview(Long id, ReviewCreateRequest review,Long reviewId){
 
-        Restaurant restaurant = restaurantRepository.findById(restaurantId).orElse(null);
-        Review oldReview = reviewRepository.findReviewByCustomerIdAndRestaurantId(id,restaurantId).orElse(null);
+        //Review oldReview = reviewRepository.findReviewByCustomerIdAndRestaurantId(id,restaurantId).orElse(null);
+        Review oldReview = reviewRepository.findById(reviewId).orElse(null);
 
         if(oldReview != null){
-            oldReview.setRestaurant(review.getRestaurant());
-            oldReview.setText(review.getText());
-            oldReview.setTitle(review.getTitle());
-            oldReview.setPoint(review.getPoint());
-
-            reviewRepository.save(oldReview);
-            return "Your review is updated!";
+            if(oldReview.getCustomer().getId() == id){
+                oldReview.setText(review.getText());
+                oldReview.setTitle(review.getTitle());
+                oldReview.setPoint(review.getPoint());
+                reviewRepository.save(oldReview);
+                return "Your review is updated!";
+            }else{
+                return "Forbidden request.";
+            }
         }else{
             return "You have not a review for this restaurant.";
         }
 
-
     }
 
-    public String deleteReview(Long id,Long restaurantId){
+    public String deleteReview(Long id,Long reviewId){
         //Customer customer = findCustomer(id);
 
-        Review review = reviewRepository.findReviewByCustomerIdAndRestaurantId(id,restaurantId).orElse(null);
+        //Review review = reviewRepository.findReviewByCustomerIdAndRestaurantId(id,restaurantId).orElse(null);
+
+        Review review = reviewRepository.findById(reviewId).orElse(null);
 
         if(review != null){
-            reviewRepository.deleteById(review.getId());
-            return "Your review was removed successfully.";
-        }
-        else{
+            if(review.getCustomer().getId() == id){
+                reviewRepository.deleteById(review.getId());
+                return "Your review was removed successfully.";
+            }else{
+                return "Forbidden request.";
+            }
+        }else{
             return "There is no such a review.";
         }
-
     }
 }
